+++ NEWS.md
@@ -5,12 +5,14 @@ nonprobsvy News and Updates
 
 ------------------------------------------------------------------------
  
-+ title corrected 
++ titles corrected 
 + new S3 method `extract` added which allows to extract results from the `nonprob` object
++ new S3 method `coef` added which allows to obtain the coefficients of underlying models (if possible)
 + fixed CRAN notes (unit tests for the IPW estimator `cloglog`) 
 + removed `sampling` package from suggested package 
 + added simple `plot` method
 + improvements in the linear algebra
++ code cleaning
 
 # nonprobsvy 0.2.0