+++ NEWS.md
@@ -12,6 +12,7 @@ nonprobsvy News and Updates
 + removed `sampling` package from suggested package 
 + added simple `plot` method
 + improvements in the linear algebra
++ corrected the `check_balance` error as in 
 + code cleaning
 
 # nonprobsvy 0.2.0